requirements.txt
supabase==2.0.0
python-dateutil==2.8.2
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
mcp>=1.0.0
openai>=1.0.0
ddgs>=1.0.0
playwright>=1.40.0
aiosqlite>=0.19.0
apscheduler>=3.10.0
